requirements.txt
transformers==4.45.2
streamlit==1.39.0
pyyaml==6.0.2